refactored smoke tests: moved common functionality to separate files, added tracking client states and final cleanup to full server test.
